CircularSuffixArray+more
Did what I think is the meat of CircularSuffixArray but am not entirely
sure.

Added some comments in MoveToFront detailing what I think should be
done.

Added our names and such in Readme.

Sadly the info needed to truly make sense of much of this will not be a
thing until tomorrow.

The teachers told me this and are themselves not prepared for this
project.

So this is sadly limited.